Split schedule into two files for clarity
